+++ src/main/java/com/json/actions/PredictionEachDataJsonAction.java
@@ -20,7 +20,7 @@ public class PredictionEachDataJsonAction extends ActionSupport {
 	private String userNameLogin;
 	private ContactPredictionBuyAsset predictionBuyAsset;
 	private String selectPrediction;
-	private float alpha;
+	private double alpha;
 	private List<String> date;
 	private List<Double> listDataCurrentYear;
 	private List<Double> listDataNextYear;
@@ -133,7 +133,8 @@ public class PredictionEachDataJsonAction extends ActionSupport {
 		this.selectPrediction = selectPrediction;
 	}
 
-	public void setAlpha(float alpha) {
+
+	public void setAlpha(double alpha) {
 		this.alpha = alpha;
 	}
